phase-v/services/websocket-sync-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
httpx==0.27.2
pydantic==2.11.7
python-dotenv==1.0.1
websockets==12.0
python-jose[cryptography]==3.3.0
